Factor fixtures out to separate file
